requirements.txt
afrimarket
requests
python-dotenv
pytz
Flask
gunicorn
setuptools